requirements.txt
python-3.9.2
pandas-1.2.3
networkx-2.5
gym-0.18.0
cvxopt-1.2.6
matplotlib-3.3.4